[cleanup] Move tests from tools/state-viewer to integration-tests directory for consistency
