+++ integration-tests/src/tests/tools/mod.rs
@@ -1,0 +1,4 @@
+mod apply_chain_range;
+mod apply_chunk;
+mod state_dump;
+mod state_viewer;
